+++ cpp/jni/com_caoccao_javet_interop_V8Native.h
@@ -233,6 +233,14 @@ JNIEXPORT jboolean JNICALL Java_com_caoccao_javet_interop_V8Native_hasOwnPropert
 
 /*
  * Class:     com_caoccao_javet_interop_V8Native
+ * Method:    idleNotificationDeadline
+ * Signature: (JJ)V
+ */
+JNIEXPORT void JNICALL Java_com_caoccao_javet_interop_V8Native_idleNotificationDeadline
+  (JNIEnv *, jobject, jlong, jlong);
+
+/*
+ * Class:     com_caoccao_javet_interop_V8Native
  * Method:    invoke
  * Signature: (JJILjava/lang/String;Z[Ljava/lang/Object;)Ljava/lang/Object;
  */